# unidad2/clase2/ejemplos/openapi-min.yaml
openapi: 3.1.1
info:
  title: Clase III · API Ejemplo
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
...
  /api/products/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    get:
      summary: Obtener producto
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404": { description: Not Found }
    patch:
      summary: Actualizar parcialmente un producto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductPatch"
      responses:
        "200": { description: OK }
        "404": { description: Not Found }
    put:
      summary: Reemplazar un producto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreate"
      responses:
        "200": { description: OK }
        "201": { description: Creado }
    delete:
      summary: Eliminar producto
      responses:
        "204": { description: No Content }
...
components:
  schemas:
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        price: { type: number }
        currency: { type: string }
        categoryId: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, name, price, currency, createdAt]
    ProductCreate:
      type: object
      properties:
        name: { type: string }
        price: { type: number }
        currency: { type: string }
        categoryId: { type: string }
      required: [name, price, currency]
    ProductPatch:
      type: object
      properties:
        name: { type: string }
        price: { type: number }
        currency: { type: string }
        categoryId: { type: string }
...